kvserver: move `SplitByLoadMergeDelay` to kvserverbase
This setting was being imported from `sql` which created
an illegal dependency.

Release note: None
