+++ pkg/kv/kvserver/kvserverbase/base.go
@@ -20,6 +20,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverpb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings"
+	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/redact"
 )
 
@@ -205,3 +206,17 @@ func IntersectSpan(
 	}
 	return
 }
+
+// SplitByLoadMergeDelay wraps "kv.range_split.by_load_merge_delay".
+var SplitByLoadMergeDelay = settings.RegisterDurationSetting(
+	"kv.range_split.by_load_merge_delay",
+	"the delay that range splits created due to load will wait before considering being merged away",
+	5*time.Minute,
+	func(v time.Duration) error {
+		const minDelay = 5 * time.Second
+		if v < minDelay {
+			return errors.Errorf("cannot be set to a value below %s", minDelay)
+		}
+		return nil
+	},
+)
