+++ pkg/kv/kvserver/replica_split_load.go
@@ -12,13 +12,11 @@ package kvserver
 
 import (
 	"context"
-	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/spanset"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
-	"github.com/cockroachdb/errors"
 )
 
 // SplitByLoadEnabled wraps "kv.range_split.by_load_enabled".
@@ -34,20 +32,6 @@ var SplitByLoadQPSThreshold = settings.RegisterIntSetting(
 	"the QPS over which, the range becomes a candidate for load based splitting",
 	2500, // 2500 req/s
 ).WithPublic()
-
-// SplitByLoadMergeDelay wraps "kv.range_split.by_load_merge_delay".
-var SplitByLoadMergeDelay = settings.RegisterDurationSetting(
-	"kv.range_split.by_load_merge_delay",
-	"the delay that range splits created due to load will wait before considering being merged away",
-	5*time.Minute,
-	func(v time.Duration) error {
-		const minDelay = 5 * time.Second
-		if v < minDelay {
-			return errors.Errorf("cannot be set to a value below %s", minDelay)
-		}
-		return nil
-	},
-)
 
 // SplitByLoadQPSThreshold returns the QPS request rate for a given replica.
 func (r *Replica) SplitByLoadQPSThreshold() float64 {
